apps/user/pyproject.toml
[project]
name = "user"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "eli", email = "eli.chandler@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastmcp>2.13",
    "pyjwt>=2.10.1",
]

[project.scripts]
user = "user:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
